Moves issue134 tests to BaseBackendTest
