+++ grid2op/tests/__init__.py
@@ -6,5 +6,8 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
-__all__ = ["BaseBackendTest",
-           "BaseRedispTest"]
+__all__ = [
+    "BaseBackendTest",
+    "BaseIssuesTest",
+    "BaseRedispTest"
+]
