+++ grid2op/tests/test_PandaPowerBackend.py
@@ -16,10 +16,16 @@ from grid2op.tests.helper_path_test import PATH_DATA_TEST_PP, PATH_DATA_TEST
 from grid2op.Backend import PandaPowerBackend
 
 from grid2op.tests.helper_path_test import HelperTests
-from grid2op.tests.BaseBackendTest import BaseTestNames, BaseTestLoadingCase, BaseTestLoadingBackendFunc
-from grid2op.tests.BaseBackendTest import BaseTestTopoAction, BaseTestEnvPerformsCorrectCascadingFailures
-from grid2op.tests.BaseBackendTest import BaseTestChangeBusAffectRightBus, BaseTestShuntAction
-from grid2op.tests.BaseBackendTest import BaseTestResetEqualsLoadGrid, BaseTestVoltageOWhenDisco, BaseTestChangeBusSlack
+from grid2op.tests.BaseBackendTest import BaseTestNames
+from grid2op.tests.BaseBackendTest import BaseTestLoadingCase
+from grid2op.tests.BaseBackendTest import BaseTestLoadingBackendFunc
+from grid2op.tests.BaseBackendTest import BaseTestTopoAction
+from grid2op.tests.BaseBackendTest import BaseTestEnvPerformsCorrectCascadingFailures
+from grid2op.tests.BaseBackendTest import BaseTestChangeBusAffectRightBus
+from grid2op.tests.BaseBackendTest import BaseTestShuntAction
+from grid2op.tests.BaseBackendTest import BaseTestResetEqualsLoadGrid
+from grid2op.tests.BaseBackendTest import BaseTestVoltageOWhenDisco
+from grid2op.tests.BaseBackendTest import BaseTestChangeBusSlack
 from grid2op.tests.BaseBackendTest import BaseIssuesTest
 PATH_DATA_TEST_INIT = PATH_DATA_TEST
 PATH_DATA_TEST = PATH_DATA_TEST_PP
